+++ fav_core/src/meta.rs
@@ -1,11 +1,11 @@
 //! Meta,
 //! making resource completely able to be operated
 
-use crate::{attr::Attr, status::Status};
+use crate::{attr::{Attr, Owner}, status::Status};
 
 /// Making resource has the complete metadata needed
 /// This is an auto trait for `T: Attr + Status`.
 /// See [`Attr`] and [`Status`].
-pub trait Meta: Attr + Status {}
+pub trait Meta: Attr + Owner + Status  {}
 
-impl<T: Attr + Status> Meta for T {}
+impl<T: Attr + Owner + Status> Meta for T {}